requirements.txt
Babel==2.4.0
blinker==1.4
click==6.7
coverage==4.4.1
decorator==4.1.2
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pbr==3.1.1
python-openid==2.2.5
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.13
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Werkzeug==0.12.2
Whoosh==2.7.4
WTForms==2.1
pyrebase=a3.0.27
mysql-python==1.2.5
